requirements.txt
fastapi[standard]==0.116.1
uvicorn[standard]
requests==2.31.0
python-multipart>=0.0.18
python-dotenv==1.0.0
pinecone-client==2.2.1
spacy>=3.7.0
dateparser>=1.1.0
cachetools==6.1.0
pydantic>=2.7.0,<3.0.0
